Add files via upload
netlist_with_buffers - It contains the spectre netlist of the architecture. Take a note of last two lines of the netlist file, they are to include verilog-A models of noise addition block and analog addition. Both of these are part of the test signal generation circuitry.\

Noise.va -  Verilog code for noise addition